+++ src/main/java/com/github/mathematicalguy/moecraft/init/ModItems.java
@@ -1,9 +1,9 @@
-package com.github.mathematicalguy.init;
+package com.github.mathematicalguy.moecraft.init;
 
 import net.minecraft.item.Item;
 import net.minecraftforge.registries.ObjectHolder;
 
-import static com.github.mathematicalguy.MoeCraftMod.MOD_ID;
+import static com.github.mathematicalguy.moecraft.MoeCraftMod.MOD_ID;
 
 @ObjectHolder(MOD_ID)
 public class ModItems {
